Revert prisma package
