+++ apps/web/src/actions/guilds.ts
@@ -3,7 +3,7 @@
 import { ActionResponse } from "@/actions";
 import { prisma } from "@albion-raid-manager/database";
 import { Guild } from "@albion-raid-manager/database/models";
-import { getServer, type Server } from "@albion-raid-manager/discord";
+import { type Server } from "@albion-raid-manager/discord";
 import logger from "@albion-raid-manager/logger";
 
 export type CreateGuildSuccessResponse = {
@@ -43,16 +43,5 @@ export async function createGuild(server: Server, userId: string) {
   } catch (error) {
     logger.error(`Failed to create guild for server ${server.id}`, error);
     return ActionResponse.Failure("CREATE_GUILD_FAILED");
-  }
-}
-
-export async function verifyServer(server: Server) {
-  try {
-    const verifiedServer = await getServer(server.id);
-    if (!verifiedServer) throw new Error("Failed to find server.");
-    return ActionResponse.Success(verifiedServer);
-  } catch (error) {
-    logger.error(`Failed to verify server ${server.name}`, { error, server });
-    return ActionResponse.Failure("SERVER_VERIFICATION_FAILED");
   }
 }
